air alarms and more minor valor tweaks (#3452)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Fixes a few outlasting issues with the Valor that I forgot from the
previous pass.
- Namely, adding air alarms and scrubbers to each room where applicable.


![image](https://github.com/user-attachments/assets/ebaba7a1-7f11-45b2-b195-13b367bd632d)

- Adding intercoms to High-Traffic Areas

- Adding Firelocks under the Cargo Blast Doors.

- Adds a Medical Webbing to the HC's Locker, since they don't spawn with
one.

Also, finally, FINALLY FIXES the WORST part of the Valor. This being the
security laptop in the way of the windoor in the armoury. THIS GAME
BREAKING FEATURE has FINALLY been fixed.


![image](https://github.com/user-attachments/assets/23f9ddbf-6c72-45f4-9024-dc74ce366c7d)


## Why It's Good For The Game
Having intercoms around the high traffic areas, hangar, surgery, mess
should be helpful. Crew on the Valor naturally congregate in the Cargo
Bay while patients generally gather outside of Surgery.
Having a full scrubber and vent network with Air Alarms is a no brainer,
again.

The firelock change was a request for better RP, as Valor crews tend to
like opening the blast doors as if they're opening the way for the
Subshuttle to land, I find this to be great fun, and I believe the Ship
Designers in-universe would put in a secondary saefty measure incase the
holofield fails. - Also people can get crushed in the doors if you trip
the fire alarm, this is funny.

## Changelog

:cl:
add: Added extra intercoms in high traffic areas on the Valor Class
add: Added Firelocks underneath the Cargo-Bay Doors on the Valor Class
add: Air Alarms, Scrubbers and Vents to every applicable room
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->