Nanotrasen Turrets (#3451)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds Nanotrasen ship turrets.

Nanotrasen ship turrets are higher rof, higher range, less damage,
somewhat slower projectiles (not a major amount, but still slower).
The only turret mapped in currently is the /light turret, on the Meta
and Delta. A medium turret, heavy turret, and pulse turret are also
defined for mapper usage.

Note: They will not actually be higher ROF until #3407 is resolved. 
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Continud turretification.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Sharplite has produced a line of turrets for Nanotrasen, which is
now mounting them on relevant vessels.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->