Makes Ration Heaters smaller. (#3486)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

Edits ration heaters to be able to be put back inside of ration packs.

It's really, really annoying that you can take ration heaters OUT of
ration packs, but CANNOT put them back in. This fixes this by making
them both the same size as the rations themselves, and adding ration
heaters to the list of accepted items that can be put in ration packs.

:cl:
code: Changes flameless ration heaters to "small" items
code: Adds flameless ration heaters to the ration pack item whitelist.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->