Resprites wisp (#3488)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
# please please, please do not port this to tg or a non-ship server.
this is me asking nicely to not to, i do not want this on a
non-shiptest-like server

## About The Pull Request


![image](https://github.com/user-attachments/assets/dc29e3a3-ede5-4d32-87db-491afe04259b)

Resprites wisp
part of project re:hair


![Screenshot_18092](https://github.com/user-attachments/assets/047f75e1-e064-47d8-9ca1-dc90a3188e2e)

## Why It's Good For The Game

hi thrax

## Changelog

:cl: rye-rice, Imaginos16
add: Resprites Wisp
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: retlaw34 <58402542+retlaw34@users.noreply.github.com>