Changes Inteq loadouts in order to add headsets to roles missing them (#3448)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This change adds basic bowmans to Enforcer and Artificer loadouts, as
well as basic headsets to Recruit loadouts.

## Why It's Good For The Game

Hearing protection is important out in the Frontier, and both Artificers
and Enforcers tend to be around things that test the limits of their
eardrums. Namely power tools and heavy machinery, or explosives and
gunfire.

## Changelog

:cl:
add: Adds bowmans to Artificer and Enforcer loadouts, and regular
headsets to Recruit loadouts
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: zimon9 <122945887+zimon9@users.noreply.github.com>
Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>