Resprites Fridges! (#3549)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Does what it says on the tin, finally implementing sprites done in
collaboration with the exceedingly talented Spookywastaken. Thank you
for that first sprite dog, I wouldn't have finished this monstrosity
without your help.

![image](https://github.com/user-attachments/assets/be838ef6-e61b-4b67-a72d-f679cd595a82)

This PR also surprisingly fixes a bug where the broken state of the
smartfridge wouldn't display due to the fact someone forgot to add it in
code. Whoops.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Death to all cruft in our server.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: Spooky, PositiveEntropy
imageadd: Smartfridges and Booze-O-Mats have been resprited!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->