Syndicate tools and soap are no longer better. (#3556)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
- syndicate soap clean speed swapped from 5 to 27 (1 better than NT
soap, because I find that funny). Also it's uses are tripled.
- "suspicious toolbox" renamed to "black and red toolbox", lost it's
damage buff.
- syndicate tools have lost the "suspicious" names, their descriptions,
and their toolspeed buff.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Cruft. Syndi items were better on stations because of rarity. On
shiptest, it feels kinda weird to map those in knowing they are just
upgrades. Soap is especially awkward in that regard.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Syndicate soap is now in-line with NT soap.
del: Syndicate tools are no longer faster. They are just reskinned
normal tools now.
del: Syndicate toolbox lost it's damage buff.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->