Reduces the chance for grenades to go off in your hands (#3441)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This change reduces the probability of a grenade blowing up in one's
hands from 15% down to 5%.

Please let me know if this probability is too low. I just chose an
oddball low number to use.

The check was also modified so that the dice roll is only made if a
projectile hits a hand holding a grenade.

## Why It's Good For The Game

It always felt a little too high of a chance to me. The probability roll
happens for _each_ projectile that one gets hit with, and with how often
and how rapidly one can be hit in a firefight here, the chance of it
blowing up in ones hands goes from being somewhat unlikely to being more
likely to happen than not.


![image](https://github.com/user-attachments/assets/4a88545b-768d-480d-b8a1-38df915e1ea2)
This calculator gives the probability of the event occurring as being
68% likely, if one were shot 7 times. This is around how many shots it
takes to bring someone down, on average, from my anecdotal experience.

This currently has the effect of making grenades useful in only a few
very specific scenarios.

If the probability of this were to go down, it would encourage the use
of grenades more, which I feel can enhance fights.

## Changelog

:cl:
balance: rebalanced the probability of a grenade going off in one's
hands
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->