Reinforces the windows on the Mudskipper (#3446)
## About The Pull Request

Compared to all the other ships of it's manufacturing company (the Kilo,
the Shetland...) the mudskipper oddly lacks reinforced windows. This
brings it up to par with the rest of the dubiously well constructed
ships adjacent to it.
Shetland:

![image](https://github.com/user-attachments/assets/935c2ff1-d90a-41c6-9afe-1a7278490dc1)
Kilo:

![image](https://github.com/user-attachments/assets/7dc153af-c149-4494-a538-059f5438b5d5)

## Why It's Good For The Game


https://github.com/user-attachments/assets/481c70a8-f2d1-460d-b2b6-947b3c18e1ea

You can also do this with a single screwdriver.

## Changelog

:cl:
fix: Mudskipper windows are correctly reinforced, like other ships by
the same manufacturer.
/:cl: