Atlas-class Light Armored Crusier (#3246)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![imagen](https://github.com/user-attachments/assets/e76c2cc9-9b25-496b-b8ff-a82b3b4debc1)

![atlas2](https://github.com/user-attachments/assets/3ca2427c-e333-4d87-8202-9f81842378b3)

The Atlas-class is the first Colonial League ship following the CLIP
Rework. The decaling and general decoration of the ship needs a
revision, but otherwise, it is intended as a small, overall generalist
patrol vessel to the likes of the Vaquero. The 4-Minutemen team, lead by
one Sergeant, is subordinate to the First Officer and Captain, who
overlook the operation of the ship by the Corpsman and the Engineer.

It is staffed by:
* 1 Captain
* 1 First Officer
* 1 Sergeant
* 1 Corpsman
* 1 Engineer
* 4 Minutemen
* 1 Deckhand

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
it's the clip ship it uses the colon league content

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Atlas-class Light Armored Crusier
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: rye-rice <58402542+rye-rice@users.noreply.github.com>
Signed-off-by: meem <75212565+meemofcourse@users.noreply.github.com>
Co-authored-by: rye-rice <58402542+rye-rice@users.noreply.github.com>