Removes the extra turret from the crying sun and slightly re-organizes the armory (#3477)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Removes an extra turret that found it's way onto the hull of the PGF
Crying Sun, likely in the recent PGF turret overhaul. Additionally, a
locker in the armory has been renamed and the flashbangs have been moved
from the Marine equipment locker to the Lieutenant's ordnance locker.
I've attached an image below to show how it's now organized.

![Screenshot 2024-10-03
172121](https://github.com/user-attachments/assets/c8c123e3-0c39-4fc6-b7e1-d2c545ac0631)
![Screenshot 2024-10-04
015626](https://github.com/user-attachments/assets/9440b65f-ff3e-48f2-8a35-05ae0f442ad6)
![Screenshot 2024-10-17
131256](https://github.com/user-attachments/assets/d2fe5685-ffb8-487a-9d2f-c7228a8ea099)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Ships having automated turrets that they weren't meant to have is
generally a bad thing for balance and being able to enjoy the ship's
lovingly labored over design. As for the flashbangs, the ship's
fragmentation and smoke grenades were already being controlled by the
lieutenant so it seems logical to do the same with grenades that can
blind and deafen someone.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
del: Removed extra turret that wasn't supposed to be on the ship
fix: Moved armory flashbangs to a more secure location
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->