resprites the commander (#3580)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
title

Was meant to be apart of another big manufacturer PR, however I do those
when i'm taking a break from big projects, so its a bit of a cycle. It
appears im in the latter half of the cycle (overmap 5 coming
december...), so might as well just PR what i have done

## Why It's Good For The Game


![image](https://github.com/user-attachments/assets/9d518d71-b8f1-487c-9070-06af2d7a6c63)

## Changelog

:cl:
add: Resprites the commander and Commissioner
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: rye-rice <58402542+rye-rice@users.noreply.github.com>
Co-authored-by: retlaw34 <58402542+retlaw34@users.noreply.github.com>
Co-authored-by: thgvr <81882910+thgvr@users.noreply.github.com>